+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes to be released in next version
 
 🙌 Improvements
  * RoomMemberDetailsVC: Enable / disable "Hide all messages from this user" from settings (#4281).
+ * RoomVC: Show / Hide More and Report Content contextual menu from settings (#4285).
 
 🐛 Bugfix
  * RoomVC: Avoid navigation to integration management using integration popup with settings set to integration disabled (#4261).